Don't set interaction ids to the input events when the scroll is active
During an active fling, a tap or a click can make it stop. We don't
want to record the user interaction UKM for the tap or click that
happens during an active fling, but still measure their input latency
and record them into performance observer. We don't set the
interaction ids for them, so we don't count them into interaction UKM.

Bug: 335238186

Change-Id: Ia2a4991d535ed342a61779ae43ed328563ae4ff0